refactor: Move bot files into a submodule
